+++ download.php
@@ -9,6 +9,7 @@ use BladeComponents\Undraw\Api\IllustrationsResponse;
 use BladeComponents\Undraw\Api\UndrawClient;
 use BladeComponents\Undraw\Api\Illustration;
 use GuzzleHttp\Client;
+use GuzzleHttp\Exception\GuzzleException;
 
 $client = new Client();
 $undrawClient = new UndrawClient();
@@ -26,12 +27,17 @@ while ($hasMore) {
     /** @var Illustration $illustration */
     foreach ($illustrationsResponse->illustrations() as $illustration) {
         $illustrationRequest = new IllustrationRequest($client);
-        $illustrationResponse = new IllustrationResponse($illustrationRequest->get($illustration->image));
+        try {
+            $illustrationResponse = new IllustrationResponse($illustrationRequest->get($illustration->image));
+        } catch (GuzzleException $e) {
+            continue;
+        }
 
         $filename = sprintf('resources/views/components/%s.blade.php', $illustration->snake());
         $svg = str_replace('fill="#6c63ff"', 'fill="{{ $color }}"', $illustrationResponse->svg());
         $svg = preg_replace('/width="\d+(.\d+)?"\s/', '', $svg, 1);
         $svg = preg_replace('/height="\d+(.\d+)?"\s/', '', $svg, 1);
+        $svg = str_replace('<?xml version="1.0" encoding="UTF-8"?>', '', $svg);
 
         file_put_contents($filename, $svg);
     }