+++ src/Api/IllustrationRequest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace BladeComponents\Undraw\Api;
 
 use GuzzleHttp\ClientInterface;
+use GuzzleHttp\Exception\GuzzleException;
 use Psr\Http\Message\ResponseInterface;
 
 /**
@@ -19,6 +20,9 @@ final class IllustrationRequest
         $this->client = $client;
     }
 
+    /**
+     * @throws GuzzleException
+     */
     public function get(string $uri): ResponseInterface
     {
         return $this->client->get($uri);